Lots of changes:
 - Now has Cmake files to build App bundle on Mac
 - Added a Cmake-enabled version of Hx at extlibs/Hx/
 - Moved source files to source/
 - Added icon
 - Added ResourcePath.hpp to get the resource path (only implemented for Mac)
 - Small changes (such as adding static casts) to files to satisfy clang

+++ source/Tilemap.cpp
@@ -72,7 +72,7 @@ void Tilemap::update()
 
 void Tilemap::setTiles(std::vector<int>& tiles)
 {
-	const unsigned int leastNumberOfTiles{ hx::min(m_tiles.size(), tiles.size()) };
+	const unsigned int leastNumberOfTiles{ static_cast<unsigned int>(hx::min(m_tiles.size(), tiles.size())) };
 	for (unsigned int t{ 0 }; t < leastNumberOfTiles; ++t)
 		m_tiles[t] = tiles[t];
 }
